Reduce kratix.io docs homepage to minimal content
This is per the GTM plan of making the current kratix.io into
docs.kratix.io and minimising the docusaurus footprint to just
docs.

When we have a proper kratix.io the complete index.tsx will
likely need to be removed, but this will remove risk of
contradictory or confusing messaging until then.